Add files via upload
Added a gear shift up function whilst in my Double Lesson today. Was fun, encountered some issues but that was due to placement of variables, specifically when current is first initialized. Need to find out how to extract single inputs when pressing a button as this can resolve some issues when shifting etc.